requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.7.1
moviepy==1.0.3
numpy==2.1.3
openai==1.54.4
pillow==11.0.0
praw==7.8.1
prawcore==2.4.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.5.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
